Extract Church encoding to a module
