+++ lambda-calculus.cabal
@@ -61,13 +61,13 @@ executable lambda-calculus
     main-is:          Main.hs
 
     -- Modules included in this executable, other than Main.
-    -- other-modules:
+    other-modules:    Church
 
     -- LANGUAGE extensions used by modules in this package.
     -- other-extensions:
 
     -- Other library packages from which modules are imported.
-    build-depends:    base ^>=4.17.2.1
+    build-depends:    base ^>=4.19.0.0
 
     -- Directories containing source files.
     hs-source-dirs:   app
